pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "caiodiasol",email = "caiodiasol@proton.me"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "django-debug-toolbar (>=4.4.6,<5.0.0)",
    "whitenoise (>=6.6.0,<7.0.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "faker (>=37.11.0,<38.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "gitpython (>=3.1.45,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]
